Fixes header split
I think that having .h instead of .hpp was confusing the compiler
